@@ -1,23 +1,23 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk.Web">
-  <PropertyGroup>
-    <TargetFramework>net5.0</TargetFramework>
-    <IncludeOpenAPIAnalyzers>true</IncludeOpenAPIAnalyzers>
-    <ReleaseVersion>1.15</ReleaseVersion>
-    <PublishWithAspNetCoreTargetManifest>false</PublishWithAspNetCoreTargetManifest>
-    <AspNetCoreHostingModel>InProcess</AspNetCoreHostingModel>
-  </PropertyGroup>
-  <ItemGroup>
-    <Folder Include="Controllers\" />
-    <Folder Include="WebHookHandlers\" />
-    <Folder Include="WebHookHandlers\Telegram\" />
-  </ItemGroup>
-  <ItemGroup>
-    <PackageReference Include="Api.Forex.Currency.Converter" Version="1.0.5" />
-    <PackageReference Include="GeoTimeZone" Version="4.1.0" />
-    <PackageReference Include="Telegram.Bot" Version="15.7.1" />
-    <PackageReference Include="Microsoft.AspNetCore.Mvc.NewtonsoftJson" Version="5.0.4" />
-  </ItemGroup>
-  <ItemGroup>
-    <None Include="appsettings.json" CopyToPublishDirectory="Always" />
-  </ItemGroup>
+    <PropertyGroup>
+        <TargetFramework>net6.0</TargetFramework>
+        <IncludeOpenAPIAnalyzers>true</IncludeOpenAPIAnalyzers>
+        <Nullable>enable</Nullable>
+        <ReleaseVersion>1.16</ReleaseVersion>
+        <DockerDefaultTargetOS>Linux</DockerDefaultTargetOS>
+    </PropertyGroup>
+    <ItemGroup>
+        <Folder Include="Controllers\" />
+        <Folder Include="WebHookHandlers\" />
+        <Folder Include="WebHookHandlers\Telegram\" />
+    </ItemGroup>
+    <ItemGroup>
+        <PackageReference Include="Api.Forex.Currency.Converter" Version="1.0.5" />
+        <PackageReference Include="GeoTimeZone" Version="4.1.0" />
+        <PackageReference Include="Telegram.Bot" Version="15.7.1" />
+        <PackageReference Include="Microsoft.AspNetCore.Mvc.NewtonsoftJson" Version="6.0.1" />
+    </ItemGroup>
+    <ItemGroup>
+        <None Include="appsettings.json" CopyToPublishDirectory="Always" />
+    </ItemGroup>
 </Project>
